Add files via upload
The Arduino code used for the control logic and output register is in these files. The programming into the EEPROMs was done with an Arduino Nano and a makeshift shift register for outputting the address of the data. The data was directly sent to the EEPROM by the Arduino Nano (see photo).